v4.0.1, v3.0.3: Skunk Memory Track Autoboot
Auto-boots the Skunk Memory Track emulation ROM
if a CD unit is detected, the user is holding down
'C' to use the CD unit, and the Skunk Memory Track
ROM is detected in bank 1. It has to be in bank 1
because it uses bank 2 to store the actual save
game data. The magic tokens used indicate a memory
track ROM are as follows:

addr:       data:
0x802004    0x534B4E4B ('SKNK')
0x802008    0x4D54524B ('MTRK')
0x80200C    boot address (w/any flags in top byte)
0x802010    Memory Track BIOS version (ignored)